pyproject.toml
[project]
name = "neo4j-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.28.2",
    "neo4j-graphrag[openai]>=1.9.0",
    "openai>=1.0,<2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
